+++ packages/core/src/handlers/getStackTrace.ts
